% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbbinom-package.R
\docType{package}
\name{cbbinom-package}
\alias{cbbinom-package}
\title{cbbinom: Continuous Analog of a Beta-Binomial Distribution}
\description{
Implementation of the d/p/q/r family of functions for a continuous analog to the standard discrete beta-binomial with continuous size parameter and continuous support with x in [0, size + 1].
}
\author{
\strong{Maintainer}: Xiurui Zhu \email{zxr6@163.com}

}
\keyword{internal}
